+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 	- Since the new syntax is defined using an object notation, individual tags can now be collapsed (#152)
 - Made `highlightColor` property optional for tags, allowing uncolored anchors
 - Omitting the path from a link anchor will now default to the current file
+- Replaced `styleComments` property with more customizeable `styleMode` property (#31)
 - Fixed epics within regions not displaying in the sidebar
 - Fixed instances where a single broken link breaks all links
 - Fixed anchors appearing outside of comments (#116)